+++ app/components/podcasts.tsx
@@ -6,19 +6,9 @@
 
 /** Dependencies. */
 import React, { useState } from 'react'
-import Link from 'next/link'
-import Image from '@/app/components/image'
-import FilterPodcasts from '@/app/components/filter-podcasts'
-
-/** Podcast Interface. */
-interface Podcast {
-  id: string
-  title: string
-  author: string
-  imageUrl: string
-  summary: string
-  link: string
-}
+import FilterPodcasts from '@/app/components/podcasts/filter'
+import ListPodcasts from '@/app/components/podcasts/list'
+import { Podcast } from '@/app/types/podcast'
 
 /** PodcastsProps Interface. */
 interface PodcastsProps {
@@ -57,32 +47,7 @@ const Podcasts: React.FC<PodcastsProps> = ({ podcasts }) => {
       />
 
       {/* Podcast List */}
-      <div className="grid grid-cols-1 sm:grid-cols-2 md:grid-cols-3 lg:grid-cols-4 gap-6">
-        {filteredPodcasts.map((podcast) => (
-          <Link
-            href={`/podcast/${podcast.id}`}
-            key={podcast.id}
-          >
-            <div className="bg-white rounded-lg shadow-lg overflow-hidden hover:shadow-xl transition-shadow duration-300 flex flex-col items-center cursor-pointer min-h-[220px]">
-              <div className="w-32 h-32 overflow-hidden rounded-full">
-                <Image
-                  src={podcast.imageUrl}
-                  alt={podcast.title}
-                  width={128}
-                  height={128}
-                  unoptimized
-                  priority
-                  className="object-cover w-full h-full"
-                />
-              </div>
-              <div className="text-center mt-2 flex-grow flex flex-col justify-center">
-                <h2 className="text-sm font-bold text-gray-600 mb-2 uppercase">{podcast.title}</h2>
-                <p className="text-sm font-semibold text-gray-400">Author: {podcast.author}</p>
-              </div>
-            </div>
-          </Link>
-        ))}
-      </div>
+      <ListPodcasts value={filteredPodcasts} />
     </div>
   )
 }
